+++ src/com/intellij/idea/plugin/hybris/project/configurators/SpringConfigurator.kt
@@ -18,13 +18,11 @@
  */
 package com.intellij.idea.plugin.hybris.project.configurators
 
-import com.intellij.idea.plugin.hybris.project.configurators.impl.DefaultSpringConfigurator
-import com.intellij.idea.plugin.hybris.project.configurators.impl.DummySpringConfigurator
 import com.intellij.idea.plugin.hybris.project.descriptors.HybrisProjectDescriptor
 import com.intellij.idea.plugin.hybris.project.descriptors.ModuleDescriptor
-import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.externalSystem.service.project.IdeModifiableModelsProvider
 import com.intellij.openapi.progress.ProgressIndicator
+import com.intellij.openapi.project.Project
 
 interface SpringConfigurator {
 
@@ -41,8 +39,6 @@ interface SpringConfigurator {
         modifiableModelsProvider: IdeModifiableModelsProvider
     )
 
-    companion object {
-        fun getInstance(): DefaultSpringConfigurator? = ApplicationManager.getApplication().getService(DefaultSpringConfigurator::class.java)
-        fun getDummyInstance(): DummySpringConfigurator = ApplicationManager.getApplication().getService(DummySpringConfigurator::class.java)
-    }
+    fun resetSpringGeneralSettings(project: Project)
+
 }